+++ src/plugins/dev-mode/unallowed-properties.ts
@@ -1,6 +1,8 @@
-import type { RxCollectionCreator } from '../../types';
+import type { RxCollectionCreator, RxDatabaseCreator } from '../../types';
 import { newRxError } from '../../rx-error';
 import { rxDatabaseProperties } from './entity-properties';
+import { validateCouchDBString } from '../../pouch-db';
+import { isFolderPath } from '../../util';
 
 
 /**
@@ -16,4 +18,28 @@ export function ensureCollectionNameValid(
             name: args.name
         });
     }
+}
+
+export function ensureDatabaseNameIsValid(args: RxDatabaseCreator) {
+
+    /**
+     * Not all strings can be used as couchdb collection name
+     * So we only allow couchdb-valid string as databse name
+     * which solves some strange bugs.
+     */
+    validateCouchDBString(args.name);
+
+    /**
+     * The server-plugin has problems when a path with and ending slash is given
+     * So we do not allow this.
+     * @link https://github.com/pubkey/rxdb/issues/2251
+     */
+    if (isFolderPath(args.name)) {
+        if (args.name.endsWith('/') || args.name.endsWith('\\')) {
+            throw newRxError('DB11', {
+                name: args.name,
+            });
+        }
+    }
+
 }